chore(npm): update dependencies
