+++ src/Helpers/DatabaseLoader.js
@@ -1,6 +1,37 @@
-import { Module, Path } from '@secjs/utils'
+import { Folder, Module, Path } from '@secjs/utils'
+import { join } from 'node:path'
 
 export class DatabaseLoader {
+  /**
+   * Return all commands from test package.
+   *
+   * @return {any[]}
+   */
+  static loadCommands() {
+    return [
+      import('#src/Commands/Db/Seed'),
+      import('#src/Commands/Db/Wipe'),
+      import('#src/Commands/Make/Model'),
+      import('#src/Commands/Make/Seeder'),
+      import('#src/Commands/Make/Resource'),
+      import('#src/Commands/Make/Migration'),
+      import('#src/Commands/Migration/Revert'),
+      import('#src/Commands/Migration/Run'),
+    ]
+  }
+
+  /**
+   * Return all custom templates from test package.
+   *
+   * @return {any[]}
+   */
+  static loadTemplates() {
+    const dirname = Module.createDirname(import.meta.url)
+    const templatesPath = join(dirname, '..', '..', 'templates')
+
+    return new Folder(templatesPath).loadSync().getFilesByPattern('**/*.ejs')
+  }
+
   /**
    * Get the schema of all models with same connection.
    *
